refactor: consolidate Aspire integration extension helpers
